LayoutTests/fast/profiler/named-functions-with-display-names-expected.txt

This page's JavaScript calls named functions and changes their display names. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. You should see functions named "[x] iteration[s] function", and not "function[x]"

Profile title: Named functions with display names
Thread_1 (no file) (line 0:0)
   startTest named-functions-with-display-names.html (line 53:19)
      0 iterations function named-functions-with-display-names.html (line 10:19)
      1 iteration function named-functions-with-display-names.html (line 17:19)
      2 iterations function named-functions-with-display-names.html (line 24:19)
      20 iterations function named-functions-with-display-names.html (line 31:20)
      1000 iterations function named-functions-with-display-names.html (line 38:22)
      bogusDisplayNameFunction named-functions-with-display-names.html (line 46:34)
      endTest profiler-test-JS-resources.js (line 1:17)
